Refactor test_award_model.py & move to tests/unit
